Introduce NPFeedback
This commit adds a new PONSO for grouping all information relating to a
single piece of feedback.

Currently it just holds a message and image, but will be expanded later
to include other meta-data.

I've also updated the handler protocol to use this feedback object, this
will allow us to later add more information without having to change the
protocol.